feat: add inital service definitions for optimizer and home store
These are two new services that will help us decide whether a home is
worth renting. Instead of just using an excel document, we can just get
any home that seems interesting to us, store it in the store, and have
the optimizer help us form the decision.

The RentalHome type is based on zillow's data for rental home features.
Some fields will not be available as some MLS services send only partial
data for a property. For the most part, zillow data should conform to
this structure.

Other listing sites will need an adapter if their data structure or
reported data is different than zillow's.