refactor: move proto definitions to common location
This follows best practices apparently. I've seen it done per service at
work which is why it was per service in this repo initially. I may need
to also rename the proto file names and subsequent imports because
apparently (from the buf docs) its bad to have the same filename for
protos in different folders. I disagree because it provides a common
convention for each of your services and makes it very easy to follow
but whatver I'm so fucking pissed that I wasted a weekend trying to get
this stupid fucking build tool to work and its just been shit the whole
time. Also the buf vscode extension constantly crashes if you make
changes to your filenames lmao good fucking tool.