feat: add firestore data model definitions for commutes
The commute service is dependent on the cached placeIds stored by the
address_wrapper service. In the future, the address wrapper service will
update the placeId in its cache based on both requests to its GetPlaceId
RPC as well as events that it subscribes to.

Google may update a place id periodically so we need to keep track of the last time a refresh
attempt was made by our service, the first time we added this address to
our cache, the time when a place id for an address was last changed, and
the last 3 most recent place ids. The historic tracking is extremely
important for other services to both update their address stores as they
may contain stale place ids that are used to associate documents in
their collections with an address/placeId

We also need to hash addresses after stripping them to a common format
(lowercase, remove whitespace, etc) so that valid address entries can be
queried regardless of formatting and return the same place id result
from the cache without needing to query from Gooogle's api.

The commute service models also model the relationship between a rental
home (to be maintained by a rental home service) and a group of points
of interest (typically the workplaces of the roommates within the
group). We duplicate the addresses so that we can more easily/quickly
query for the computed commute times for a group and return the basic
relationships.

[Commute][AddressWrapper]