refactor: move point of interest store
It makes more sense intuitively to store each user's points of interest
with the roommate service instead of the commute service. The commute
service will still hold a replica of the user's points of interest used
in a commute computation in its own store but management of these POIs
will be done through the roommate service.

[Commute][Roommate]