feat: move add point of interest rpc to roommate service
The roommate service is responsible for handling the users of the
platform. This includes management of each user's particular points of
interest. Points of interest should be added to groups and the user
should be able to create points of interest as they need.

The Commute service should only be responsible for computing commute
times and routes for a rental address and all of the points of interest
within a group.

[Commute][Roommate]