feat: initial commute service implementation
This creates a framework for the commute service. It uses grpc and
connects to google's routes api, a roommate service, and address wrapper
service, and a firestore db to store computed commutes for points of
interest within a roommate group and a rental home.

It currently only supports finding commute times during typical working
months with the heaviest traffic. It excludes calculating commutes
during December and January as well as the summer months. This is
typically when people are on vacation and there are fewer people
commuting in the morning. It also uses 7:00am CST/CDT on Thursday as the
day of the commute.

Right now it tries to calculate the commute 2 months in the future, when
google will use more historical data. If this date falls within the
excluded months, then it will try to find the next month that isn't
excluded.

This still requires testing but theoretically it should work™.

[Commute]