feat: add create point of interest and get group pois RPCs
AddPointOfInterest was renamed to CreatePointOfInterest. Instead
AddPointOfInterest will add a user's point of interest to a group. A
user can create any number of points of interest for themselves and has
the option to include it in their group. A user should only be allowed
to add at maximum 5 points of interest since there are at most 8 active
members of a group.

The GetGroupPointsOfInterest RPC is used internally by the commute
service and should not be exposed outside of gRPC. It sends the points
of interest within a group to the calling service. For commute, this is
used to form the ComputeRouteMatrix request to google's routes api.

[Roommate]