feat: add roommate service implmentation
This creates the basic roommate service skeleton. No validation is or
error messages are implemented but for now we can easily use this
service with the commute service. Data is updated and created in a
deployed firestore db.

Roommates (users) can be created, roommates can create groups and points of
interest and can add their points of interest to a group

[Roommate]