fix: infinite loop when creating commute request and seg faults
Commute service would segfault when the request builder tried to make a
request. This is because the protocol buffer fields weren't being
initalized correctly (you cant use dot notation like I was doing)
causing a segfault when an inner deeply nested field was initialized.

There was also an inifinite loop happening when the builder tried to get
a month that was outside of the excluded bounds and a time on thursday.
This was because I didn't set the updated date back to the variable.

I also fixed an error where adding multiple POIs to a group would
overwrite existing POIs so there would always be one. This happened
because I also didn't initalize a list correctly to use the existing
POIs in the group. These were all dumb mistakes lol

[Commute][Roommate]