ci: add basic build and deploy job
This just pushes an image to a docker repo that's private. It doesn't deploy to a
cloud run deployment yet. It attempts to only push new docker images for
the changed services from the commit hash as
long as the infrastructure folder isn't changed since this shouldn't
trigger a new image build. Infra should be managed by another workflow.

[CI]