chore: provision artifact registry and github ci service account
I need artifact registry to push and pull the container images for the services. I originally was going to use docker hub with the 1 free private repository that you get as a free user but for this project it makes more sense to stick within the GCP ecosystem since the cost will be relatively minimal for the expected number of images/services. Also I would need to connect docker hub with GCP anyway so it would have been a massive pain regardless.

I forgot about federated workload identities so in the near future, I'll
have to remove the service account and provision a workload identity
pool instead. Service accounts are less secure because you need to store
a key.json file securely (which is hard to do).

[Infra]