chore: provision WIF pool and provider
For CICD, the recommended process is to directly bind a Workload
Identity Federation Pool principal/principalSet to a role, as long as
the applicable service supports it. Artifact registry does so Instead of
Service Account impersonation I'm using a direct binding. In the future,
I'll probably end up using SA impersonation but this is what I saw based
on the docs for the google-github-actions@v3: https://github.com/google-github-actions/auth?tab=readme-ov-file#direct-wif

[Infra]