chore: build and push container image to gcp
This changes the build-deploy action to deploy to GCP via Artifact
Registry instead of to docker hub. I'm already stuck using GCP for the
rest of my services so makes sense to just continue in this environment.

Instead of doing a multistage build in a Dockerfile, it's easier (for
now) to build outside of docker and copy the resulting build artifact
into the docker container. This is due to the monorepo structure I have
with a toplevel go.work and each service having a separate go module.
Resolving dependencies will be a pain to do in a Dockerfile because of
the Dockerfile context, so this is the easy way out.

[AddressWrapper][CICD]