fix: detect changes for changed services
Moved the detection logic to a separate script for improved
maintainability. I was not updating the arrays correctly and using the
incorrect bash variables.

I also removed the filter for changes to the infrastructure subdirectory
for now since this complicated the logic. Will need to add a filter in
the future that is more modular.

[CICD]