chore: ignore folders and files from triggering build
Certain files and folders like the infrastructure and README don't need
to trigger new builds. There is nothing new about the code when only
these files change and creating a new build would be redundant and add
cost. If infra is provisioned AND there's a code change then a new build
should still be triggered.

[CICD]