Revert "fix: temporarily allow all traffic on dev address wrapper service"
This reverts commit 5d147d1dd47b81b3590774e97e763c98bc323add.

The service is working and I'm able to get the place id of a location.
This means that I'll need to add the service accounts for various
invoker services or use mTLS to call once the other services are
deployed. Right now, it seems difficult to test a single service locally
that depends on Address Wrapper and makes upstream calls to it due to
the ingress configuration.

[AddressWrapper]