fix: invalid listener host
net.Listen() expects the address to listen on to be of the form
"ADDRESS:PORT". This wasn't configured properly on the startup causing
the container to fail. Since these services are hosted on Cloud Run we
can just call the net.JoinHostPort() function with an empty address and
it will automatically listen on the configured port on 127.0.0.1.

[Commute][Roommate]