chore: deploy commute service
This provisions a new cloud run service for the commute service. It is
configured to be an internal service and requires an envoy sidecar or an
envoy edge proxy before it can be used. This is because frontends cannot
call this service without a json -> gRPC transcoding and this service is
expected to be customer-facing.

[Commute]