chore: automate envoy sidecar image build step
Each service should have its own Envoy sidecar if it's a public-facing
service. This primarily applies to the commute service since it's the
primary driver of the project. Instead of manually building this, I just
want to be able to push my chagnes and have the CI publish the image to
google artifact registry so I can update the terraform files with the
appropriate version. This way I can add more Envoy sidecars as the
project increases in scope.

[CICD]