chore: create envoy sidecar for commute service
This sidecar needs to both authenticate incoming requests and provide
JSON-gRPC transcoding since the Commute Service is public-facing.
Authentication is provided by Auth0 rather than GCP since it will not
add any cost. It increases the complexity and I could have probably just
used a service account and authenticated that way instead of doing this
but it was worth it to start working with Envoy.

[Commute]