fix: invalid property supplied to envoy configuration
The envoy configuration defines the lb_endpoints proto as containing a
list of endpoint messages. Each endpoint message contains an address
message (see v3 configuration documentation for more information). The
configuration I made was missing an extra tab space for each address
configuration to add it as a message under the endpoint message above,
causing the startup to fail.

[Commute][Envoy]