chore: deploy envoy sidecar
This adds an envoy sidecar to the commute service. The commute service
will allow external traffic so I need to transcode the JSON -> gRPC. I
also need to authenticate requests so that if request traffic spikes,
the costs remain low and other services are not called (which may be
billed for gcp apis or other upstream services they interact with).

[Commute]