fix: make services public
All services are accessible via the public internet but still require
IAM permissions to be invoked. This is more favourable at the moment
since it's easier to configure it this way.

The alternative requires a VPC network to be configured which will take
a significant amount of effort to do and will make this project more
costly. I'd rather not do this yet and want a functioning POC before
resorting to using a VPC.

[AddressWrapper][Roommate]