fix: require authentication for all routes
The commute service is public-facing but only clients with the correct
access credentials from auth0 should be able to make requests to the
service. This is because the commute service invokes two other private
services and having authentication would limit the risk of malformed
data making its way into the datastores controlled by these private
services.

The envoy config wasn't setup correctly and was missing the rules field.

[Commute][Envoy]