chore: add logging for public services
Logs currently aren't present aside from receiving a request in GCP. I
need observability into what the commute service is doing and the kinds
of requests its sending to other services. I also need to see how its
processing the response as well as the response it's returning.

This is only partial observability since there's no span or traceId
available to see service-to-service communication.

This adds a wrapper struct around slog with a global singleton. This
allows the logger to be available for all services separately without
needing to pass the logger to functions or other structs. It wraps
around slog so that it can be customized in the future as needed,
including switching to a different logger altogether without affecting
consumers of the logger.

[Commute]