fix: create geocode request with all structured parameters
The address_wrapper service was returning placeIds for a completely
different location than the one specified in the address message of the
request. This was because I was relying only on the address and
countryCode which were the minimum structured fields required for to
geocode an address.

Instead, all fields should be added to the geocode request. This may
cause issues for countries without supported fields like
administrativeArea or postalCode but in most cases this will return the
correct placeId.

[AddressWrapper]