feat: calculate return commute time
This allows users to supply a departure time from the home to get to all
points of interest as well as a departure time from all points of
interest to get back to the home. Both of these options can be specified
(home_departure_time, point_of_interest_departure_time). The former typically
corresponds with commuting to work and the later corresponds with
commuting from work.

If an option is not specified, home_departure_time defaults to 7:00am
CST and point_of_interest_departure_time defaults to 4:30pm CST.

[Commute]