chore: triggering a new build of envoy image
Need to trigger a new build of the envoy image since the proto
descriptors have been updated for the commute service in
37046c3f2a34d298d0921b24d93cf9b0df24867e

Nothing here is really changed just adding an extra line for this new
build. This fixes an issue where the response is coming with a 200
success code but an empty object as a response body.

[Commute]