fix: allow empty token source for local env
For local development, it's not possible to start the server locally
when running `make run-local` because we need an id token from gcp.
Instead of attempting to request this token (which is just a jwt in
reality), the authentication process should be skipped entirely if the
configured upstream service is hosted on the loopback address.

This makes sure that services hosted locally (like the address wrapper)
don't require auth and can be developed easily in conjunction with the
commute service.

[Commute]